refactor: infra 패키지에 있는 파일 이동
