+++ src/main/java/gg/agit/konect/domain/upload/service/UploadService.java
@@ -12,8 +12,8 @@ import org.springframework.web.multipart.MultipartFile;
 import gg.agit.konect.domain.upload.dto.ImageUploadResponse;
 import gg.agit.konect.global.code.ApiResponseCode;
 import gg.agit.konect.global.exception.CustomException;
-import gg.agit.konect.infra.storage.cdn.StorageCdnProperties;
-import gg.agit.konect.infra.storage.s3.S3StorageProperties;
+import gg.agit.konect.infrastructure.storage.cdn.StorageCdnProperties;
+import gg.agit.konect.infrastructure.storage.s3.S3StorageProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import software.amazon.awssdk.core.exception.SdkClientException;
