+++ src/main/java/gg/agit/konect/global/config/SecurityConfig.java
@@ -21,9 +21,9 @@ import org.springframework.security.web.SecurityFilterChain;
 import gg.agit.konect.global.auth.oauth.OAuth2LoginSuccessHandler;
 import gg.agit.konect.global.auth.oauth.OAuth2RedirectUriSaveFilter;
 import gg.agit.konect.global.auth.oauth.SocialOAuthService;
-import gg.agit.konect.infra.oauth.AppleClientSecretProvider;
-import gg.agit.konect.infra.oauth.AppleOAuthServiceImpl;
-import gg.agit.konect.infra.oauth.CustomRequestEntityConverter;
+import gg.agit.konect.infrastructure.oauth.AppleClientSecretProvider;
+import gg.agit.konect.infrastructure.oauth.AppleOAuthServiceImpl;
+import gg.agit.konect.infrastructure.oauth.CustomRequestEntityConverter;
 import lombok.RequiredArgsConstructor;
 
 @Configuration
