+++ src/main/java/gg/agit/konect/infrastructure/oauth/AppleOAuthProperties.java
@@ -1,4 +1,4 @@
-package gg.agit.konect.infra.oauth;
+package gg.agit.konect.infrastructure.oauth;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
