+++ src/main/java/gg/agit/konect/infrastructure/oauth/AppleOAuthServiceImpl.java
@@ -1,4 +1,4 @@
-package gg.agit.konect.infra.oauth;
+package gg.agit.konect.infrastructure.oauth;
 
 import java.util.Optional;
