+++ src/main/java/gg/agit/konect/infrastructure/oauth/CustomRequestEntityConverter.java
@@ -1,4 +1,4 @@
-package gg.agit.konect.infra.oauth;
+package gg.agit.konect.infrastructure.oauth;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.oauth2.client.endpoint.DefaultOAuth2TokenRequestParametersConverter;
