+++ src/main/java/gg/agit/konect/infrastructure/storage/cdn/StorageCdnProperties.java
@@ -1,4 +1,4 @@
-package gg.agit.konect.infra.storage.cdn;
+package gg.agit.konect.infrastructure.storage.cdn;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
