+++ src/main/java/gg/agit/konect/infrastructure/storage/s3/S3StorageProperties.java
@@ -1,4 +1,4 @@
-package gg.agit.konect.infra.storage.s3;
+package gg.agit.konect.infrastructure.storage.s3;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
